Move shared logic to abstract base class
