+++ packages/framework/src/Foundation/RouteCollection.php
@@ -2,11 +2,9 @@
 
 namespace Hyde\Framework\Foundation;
 
-use Hyde\Framework\Contracts\HydeKernelContract;
 use Hyde\Framework\Contracts\PageContract;
 use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Models\Route;
-use Illuminate\Support\Collection;
 
 /**
  * Pseudo-Router for Hyde.
@@ -30,27 +28,8 @@ use Illuminate\Support\Collection;
  * determine where a source file will be compiled to, and where a compiled
  * file was generated from.
  */
-final class RouteCollection extends Collection
+final class RouteCollection extends BaseSystemCollection
 {
-    protected HydeKernelContract $kernel;
-
-    public static function boot(HydeKernelContract $kernel): self
-    {
-        return (new self())->setKernel($kernel)->runDiscovery();
-    }
-
-    protected function __construct($items = [])
-    {
-        parent::__construct($items);
-    }
-
-    protected function setKernel(HydeKernelContract $kernel): self
-    {
-        $this->kernel = $kernel;
-
-        return $this;
-    }
-
     public function getRoutes(?string $pageClass = null): self
     {
         return ! $pageClass ? $this : $this->filter(function (RouteContract $route) use ($pageClass) {
